+++ ApexGameEngine/src/Apex/Graphics/Renderer/RendererAPI.h
@@ -66,7 +66,7 @@ namespace Apex {
 
 		virtual void Draw(const Ref<VertexArray>& vertexArray) = 0;
 		virtual void DrawIndexed(const Ref<VertexArray>& vertexArray, uint32_t indexCount = 0) = 0;
-		virtual void DrawInstanced(const Ref<VertexArray>& vertexArray, size_t count) = 0;
+		virtual void DrawInstanced(const Ref<VertexArray>& vertexArray, uint32_t count) = 0;
 
 		virtual void SetDepthTest(bool value) = 0;
 		virtual void SetDepthWriteMode(bool value) = 0;
